fix: voice message

+++ ipbandroidview/src/main/kotlin/com/progressterra/ipbandroidview/core/voice/AudioManager.kt
@@ -44,6 +44,6 @@ interface AudioManager {
         }
 
         override fun progress(): Float =
-            (mediaPlayer.currentPosition / mediaPlayer.duration).toFloat()
+            mediaPlayer.currentPosition.toFloat() / mediaPlayer.duration.toFloat()
     }
 }
